archetypes/api-service/requirements.txt
# Core FastAPI
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
pydantic>=2.0.0
pydantic-settings>=2.0.0

# Authentication & Security
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.9

# Rate Limiting & Caching
redis>=5.0.0
slowapi>=0.1.9

# Monitoring
prometheus-fastapi-instrumentator>=7.0.0

# Database & Migrations
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.10
alembic>=1.13.3

# Development
pytest>=8.0.0
pytest-asyncio>=0.23.0
httpx>=0.27.0
black>=24.0.0
pylint>=3.0.0
